cli: Removing negative value check for unsigned variable (coverity fix)
CID: 1124663

BUG: 789278
Reviewed-on: http://review.gluster.org/9669
Tested-by: Gluster Build System <jenkins@build.gluster.com>

+++ cli/src/cli-cmd-parser.c
@@ -2582,7 +2582,7 @@ cli_cmd_volume_top_parse (const char **words, int wordcount,
         int32_t  list_cnt       = -1;
         int      index          = 0;
         int      perf           = 0;
-        uint32_t  blk_size      = 0;
+        int32_t  blk_size      = 0;
         uint32_t  count         = 0;
         gf_boolean_t nfs        = _gf_false;
         char    *delimiter      = NULL;
@@ -2703,7 +2703,8 @@ cli_cmd_volume_top_parse (const char **words, int wordcount,
                                 ret = -1;
                                 goto out;
                         }
-                        ret = dict_set_uint32 (dict, "blk-size", blk_size);
+                        ret = dict_set_uint32 (dict, "blk-size",
+                                                        (uint32_t)blk_size);
                 } else if (perf && !nfs && !strcmp (key, "count")) {
                         ret = gf_is_str_int (value);
                         if (!ret)
